create unittest._log stub
